requirements.txt
APScheduler==3.10.4
backports.zoneinfo;python_version<"3.9"
blinker==1.7.0
cachelib==0.10.2
cffi==1.16.0
click==8.1.7
confluent-kafka==2.3.0
cryptography==41.0.7
Flask==3.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzlocal==5.2
Werkzeug==3.0.1
zipp==3.17.0
